Updating window dimensions on resize breaks page
Calling the `update_window_dimensions` function on `window.resize` breaks the page, as on line 83 it does `  	jQuery('#wpbody').css('height', body_height).css('overflow', 'hidden');`.  So whenever a user resizes the window, the `#wpbody` will end up with overflow hidden, effectively breaking the page for the user.

There's no need to attach an event listener to the window resize since there's already a click event on the cloudinary button.